+++ AppLinking/Core/AssemblyInfo.cs
@@ -1,7 +1,7 @@
 using System.Runtime.CompilerServices;
 using Unity.Cloud.Common;
 
-[assembly: ApiSourceVersion("com.unity.cloud.app-linking", "1.1.4")]
+[assembly: ApiSourceVersion("com.unity.cloud.app-linking", "1.1.5")]
 #if !(UC_NUGET)
 [assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Editor")]
 [assembly: InternalsVisibleTo("Unity.Cloud.AppLinking.Tests")]